+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- Explicit directory registration host properly taken into account by model registry, trial datastore and web proxy
+- Failure of host self discovery are now reported
+
 ## v2.17.4 - 2023-08-21
 
 ### Fixed